dockerfiles/benchmarks/requirements.txt
prometheus_api_client==0.4.2
PyGithub==1.55
